+++ appnav/src/main/kotlin/io/element/android/appnav/di/MatrixClientsHolder.kt
@@ -16,69 +16,3 @@
 
 package io.element.android.appnav.di
 
-import com.bumble.appyx.core.state.MutableSavedStateMap
-import com.bumble.appyx.core.state.SavedStateMap
-import io.element.android.libraries.matrix.api.MatrixClient
-import io.element.android.libraries.matrix.api.auth.MatrixAuthenticationService
-import io.element.android.libraries.matrix.api.core.SessionId
-import kotlinx.coroutines.runBlocking
-import timber.log.Timber
-import java.util.concurrent.ConcurrentHashMap
-import javax.inject.Inject
-
-private const val SAVE_INSTANCE_KEY = "io.element.android.x.di.MatrixClientsHolder.SaveInstanceKey"
-
-class MatrixClientsHolder @Inject constructor(private val authenticationService: MatrixAuthenticationService) {
-
-    private val sessionIdsToMatrixClient = ConcurrentHashMap<SessionId, MatrixClient>()
-
-    fun add(matrixClient: MatrixClient) {
-        sessionIdsToMatrixClient[matrixClient.sessionId] = matrixClient
-    }
-
-    fun removeAll() {
-        sessionIdsToMatrixClient.clear()
-    }
-
-    fun remove(sessionId: SessionId) {
-        sessionIdsToMatrixClient.remove(sessionId)
-    }
-
-    fun isEmpty(): Boolean = sessionIdsToMatrixClient.isEmpty()
-
-    fun knowSession(sessionId: SessionId): Boolean = sessionIdsToMatrixClient.containsKey(sessionId)
-
-    fun getOrNull(sessionId: SessionId): MatrixClient? {
-        return sessionIdsToMatrixClient[sessionId]
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    fun restore(state: SavedStateMap?) {
-        Timber.d("Restore state")
-        if (state == null || sessionIdsToMatrixClient.isNotEmpty()) return Unit.also {
-            Timber.w("Restore with non-empty map")
-        }
-        val sessionIds = state[SAVE_INSTANCE_KEY] as? Array<SessionId>
-        Timber.d("Restore matrix session keys = ${sessionIds?.map { it.value }}")
-        if (sessionIds.isNullOrEmpty()) return
-        // Not ideal but should only happens in case of process recreation. This ensure we restore all the active sessions before restoring the node graphs.
-        runBlocking {
-            sessionIds.forEach { sessionId ->
-                Timber.d("Restore matrix session: $sessionId")
-                authenticationService.restoreSession(sessionId)
-                    .onSuccess { matrixClient ->
-                        add(matrixClient)
-                    }
-                    .onFailure {
-                        Timber.e("Fail to restore session")
-                    }
-            }
-        }
-    }
-
-    fun save(state: MutableSavedStateMap) {
-        val sessionKeys = sessionIdsToMatrixClient.keys.toTypedArray()
-        Timber.d("Save matrix session keys = ${sessionKeys.map { it.value }}")
-        state[SAVE_INSTANCE_KEY] = sessionKeys
-    }
-}
