+++ changelog.d/880.bugfix
@@ -1,0 +1,1 @@
+Fix sliding sync loop restarts due to expirations.
